+++ module/HomebrewMacros.js
@@ -21,7 +21,7 @@ class HomebrewMacros {
                 //wait for initial render
                 await warpgate.wait(100);
                 const ray = new Ray( tokenCenter, crosshairs );
-                const distance = canvas.grid.measureDistances([{ray}], {gridSpaces:true})[0]
+                const distance = canvas.grid.measureDistances([{ray}], {gridSpaces:true})[0];
 
                 //only update if the distance has changed
                 if (cachedDistance !== distance) {
@@ -882,10 +882,10 @@ class HomebrewMacros {
         }
 
         let squares = maxSquares ? maxSquares : 1;
-        let pixelsPerSquare = canvas.grid.size * 1.33; // handle diagonals
-        let movePixels = squares * pixelsPerSquare;
+        let pullBackFt = 5 * squares;
+        let pullBackFactor = pullBackFt / canvas.dimensions.distance;
         const ray = new Ray(pullerToken.center, targetToken.center);
-        let newCenter = ray.project((ray.distance - movePixels)/ray.distance);
+        let newCenter = ray.project(1 - ((canvas.dimensions.size * pullBackFactor) / ray.distance));
 
         // check for collision
         let c = canvas.grid.getSnappedPosition(newCenter.x - targetToken.width / 2, newCenter.y - targetToken.height / 2, 1);
@@ -893,8 +893,10 @@ class HomebrewMacros {
 
         while ((squares > 1) && !isAllowedLocation) {
             squares = squares - 1;
-            movePixels = squares * pixelsPerSquare;
-            let shorterCenter = ray.project((ray.distance - movePixels)/ray.distance);
+            pullBackFt = 5 * squares;
+            pullBackFactor = pullBackFt / canvas.dimensions.distance;
+
+            let shorterCenter = ray.project(1 - ((canvas.dimensions.size * pullBackFactor) / ray.distance));
             c = canvas.grid.getSnappedPosition(shorterCenter.x - targetToken.width / 2, shorterCenter.y - targetToken.height / 2, 1);
             let isShorterAllowed = !HomebrewMacros.checkPosition(c.x, c.y);
 
@@ -916,6 +918,10 @@ class HomebrewMacros {
         return false;
     }
 
+    static round5(x) {
+        return (x % 5) >= 2.5 ? parseInt(x / 5) * 5 + 5 : parseInt(x / 5) * 5;
+    }
+
     /**
      * Pushes the target maxSquares number of squares away from the pusher.
      *
@@ -932,10 +938,10 @@ class HomebrewMacros {
         }
 
         let squares = maxSquares ? maxSquares : 1;
-        let pixelsPerSquare = canvas.grid.size * 1.33; // handle diagonals
-        let movePixels = squares * pixelsPerSquare;
+        let knockBackFt = 5 * squares;
+        let knockBackFactor = knockBackFt / canvas.dimensions.distance;
         const ray = new Ray(pusherToken.center, targetToken.center);
-        let newCenter = ray.project((ray.distance + movePixels)/ray.distance);
+        let newCenter = ray.project(1 + ((canvas.dimensions.size * knockBackFactor) / ray.distance));
 
         // check for collision
         let c = canvas.grid.getSnappedPosition(newCenter.x - targetToken.width / 2, newCenter.y - targetToken.height / 2, 1);
@@ -943,8 +949,11 @@ class HomebrewMacros {
 
         while ((squares > 1) && !isAllowedLocation) {
             squares = squares - 1;
-            movePixels = squares * pixelsPerSquare;
-            let shorterCenter = ray.project((ray.distance + movePixels)/ray.distance);
+            knockBackFt = 5 * squares;
+            knockBackFactor = knockBackFt / canvas.dimensions.distance;
+
+            //movePixels = squares * pixelsPerSquare;
+            let shorterCenter = ray.project(1 + ((canvas.dimensions.size * knockBackFactor) / ray.distance));
             c = canvas.grid.getSnappedPosition(shorterCenter.x - targetToken.width / 2, shorterCenter.y - targetToken.height / 2, 1);
             let isShorterAllowed = !HomebrewMacros.checkPosition(c.x, c.y);